Prefer C++ method std::ceil over C method ceil

+++ gpu/people/src/cuda/nvidia/NPP_staging.cu
@@ -2366,12 +2366,12 @@ __global__ void resizeSuperSample_32f(NcvSize32u srcSize,
     float yEnd   = fmin (y + scaleY, rh - 1.0f);
     // x range of source samples
     float floorXBegin = floorf (xBegin);
-    float ceilXEnd    = ceilf (xEnd);
+    float ceilXEnd    = std::ceil (xEnd);
     int iXBegin = srcROI.x + (int) floorXBegin;
     int iXEnd   = srcROI.x + (int) ceilXEnd;
     // y range of source samples
     float floorYBegin = floorf (yBegin);
-    float ceilYEnd    = ceilf (yEnd);
+    float ceilYEnd    = std::ceil (yEnd);
     int iYBegin = srcROI.y + (int) floorYBegin;
     int iYEnd   = srcROI.y + (int) ceilYEnd;
 
@@ -2451,10 +2451,10 @@ __global__ void resizeBicubic(NcvSize32u srcSize,
 
     // sampling range
     // border mode is clamp
-    float xmin = fmax (ceilf (x - 2.0f), 0.0f);
+    float xmin = fmax (std::ceil (x - 2.0f), 0.0f);
     float xmax = fmin (floorf (x + 2.0f), rw - 1.0f);
 
-    float ymin = fmax (ceilf (y - 2.0f), 0.0f);
+    float ymin = fmax (std::ceil (y - 2.0f), 0.0f);
     float ymax = fmin (floorf (y + 2.0f), rh - 1.0f);
 
     // shift data window to match ROI
